requirements.txt
# Core
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Cors==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.21
geoalchemy2==0.14.7

# Data Processing
pandas==2.1.0
numpy==1.25.2
geopandas==0.13.2
shapely==2.0.1
geopy==2.4.0

# Map Visualization
folium==0.15.1
branca==0.8.1

# Authentication
email-validator==2.0.0
Flask-Bcrypt==1.0.1
itsdangerous==2.1.2

# API
Flask-RESTful==0.3.10
marshmallow==3.20.1
marshmallow-sqlalchemy==1.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
python-slugify==8.0.1

# Development
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
isort==5.12.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
folium==0.20.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langsmith==0.4.4
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.3.1
openai==1.93.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xyzservices==2025.4.0
zstandard==0.23.0
